feat: fixed dependency management and moved common classes to new library
