+++ src/main/java/at/adesso/leagueapi/gamedataservice/application/datadragon/summonerspells/SummonerSpellApplicationService.java
@@ -1,9 +1,9 @@
 package at.adesso.leagueapi.gamedataservice.application.datadragon.summonerspells;
 
+import at.adesso.leagueapi.commons.errorhandling.exception.ResourceNotFoundException;
 import at.adesso.leagueapi.gamedataservice.application.datadragon.AbstractDataDragonApplicationService;
 import at.adesso.leagueapi.gamedataservice.domain.summonerspells.SummonerSpell;
 import at.adesso.leagueapi.gamedataservice.infrastructure.adapter.riot.summonerspells.RiotSummonerSpellAdapter;
-import at.adesso.leagueapi.gamedataservice.infrastructure.config.errorhandling.exception.ResourceNotFoundException;
 import at.adesso.leagueapi.gamedataservice.infrastructure.config.model.DataDragonProperties;
 import org.springframework.stereotype.Service;
