+++ src/main/java/at/adesso/leagueapi/gamedataservice/application/summoners/SummonerApplicationService.java
@@ -1,8 +1,8 @@
 package at.adesso.leagueapi.gamedataservice.application.summoners;
 
+import at.adesso.leagueapi.commons.errorhandling.exception.ResourceNotFoundException;
 import at.adesso.leagueapi.gamedataservice.domain.accounts.model.Summoner;
 import at.adesso.leagueapi.gamedataservice.infrastructure.adapter.riot.summoners.RiotSummonerAdapter;
-import at.adesso.leagueapi.gamedataservice.infrastructure.config.errorhandling.exception.ResourceNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
