+++ src/main/java/at/adesso/leagueapi/gamedataservice/infrastructure/adapter/riot/summoners/mapper/SummonerApiDtoMapper.java
@@ -1,8 +1,8 @@
 package at.adesso.leagueapi.gamedataservice.infrastructure.adapter.riot.summoners.mapper;
 
+import at.adesso.leagueapi.commons.mapper.DefaultMapperConfig;
 import at.adesso.leagueapi.gamedataservice.domain.accounts.model.Summoner;
 import at.adesso.leagueapi.gamedataservice.infrastructure.adapter.riot.summoners.model.SummonerApiDto;
-import at.adesso.leagueapi.gamedataservice.infrastructure.config.mapper.DefaultMapperConfig;
 import org.mapstruct.Mapper;
 
 @Mapper(config = DefaultMapperConfig.class)
