+++ src/main/java/at/adesso/leagueapi/gamedataservice/infrastructure/adapter/riot/summonerspells/mapper/SummonerSpellDtoMapper.java
@@ -1,8 +1,8 @@
 package at.adesso.leagueapi.gamedataservice.infrastructure.adapter.riot.summonerspells.mapper;
 
+import at.adesso.leagueapi.commons.mapper.DefaultMapperConfig;
 import at.adesso.leagueapi.gamedataservice.domain.summonerspells.SummonerSpell;
 import at.adesso.leagueapi.gamedataservice.infrastructure.adapter.riot.summonerspells.model.SummonerSpellDto;
-import at.adesso.leagueapi.gamedataservice.infrastructure.config.mapper.DefaultMapperConfig;
 import org.mapstruct.Mapper;
 
 import java.util.List;
