+++ src/main/java/at/adesso/leagueapi/gamedataservice/infrastructure/config/errorhandling/ExceptionHandlerAdvice.java
@@ -1,16 +1,14 @@
-package at.adesso.leagueapi.gamedataservice.infrastructure.api.rest.common;
+package at.adesso.leagueapi.gamedataservice.infrastructure.config.errorhandling;
 
 import at.adesso.leagueapi.commons.errorhandling.AbstractExceptionHandlerAdvice;
 import at.adesso.leagueapi.commons.errorhandling.ErrorResponseFactory;
 import lombok.extern.slf4j.Slf4j;
 import org.slf4j.Logger;
-import org.springframework.web.bind.annotation.ControllerAdvice;
 
 @Slf4j
-@ControllerAdvice
 public class ExceptionHandlerAdvice extends AbstractExceptionHandlerAdvice {
 
-    protected ExceptionHandlerAdvice(final ErrorResponseFactory responseFactory) {
+    protected ExceptionHandlerAdvice(ErrorResponseFactory responseFactory) {
         super(responseFactory);
     }
