--- src/main/java/at/adesso/leagueapi/gamedataservice/infrastructure/config/errorhandling/JsonErrorController.java
@@ -1,27 +1,0 @@
-package at.adesso.leagueapi.gamedataservice.infrastructure.config.errorhandling;
-
-import at.adesso.leagueapi.gamedataservice.infrastructure.config.errorhandling.exception.ResourceNotFoundException;
-import at.adesso.leagueapi.gamedataservice.infrastructure.config.errorhandling.exception.TechnicalException;
-import jakarta.servlet.RequestDispatcher;
-import jakarta.servlet.http.HttpServletRequest;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.GetMapping;
-
-public class JsonErrorController {
-
-    @GetMapping("/error")
-    public void handleError(final HttpServletRequest request) {
-        final Object status = request.getAttribute(RequestDispatcher.ERROR_STATUS_CODE);
-
-        if (status != null) {
-            final int statusCode = Integer.parseInt(status.toString());
-
-            if (statusCode == HttpStatus.NOT_FOUND.value()) {
-                throw new ResourceNotFoundException();
-            }
-        }
-
-        // Should actually not happen
-        throw new TechnicalException();
-    }
-}
