--- src/main/java/at/adesso/leagueapi/gamedataservice/infrastructure/config/errorhandling/model/ErrorResponse.java
@@ -1,16 +1,0 @@
-package at.adesso.leagueapi.gamedataservice.infrastructure.config.errorhandling.model;
-
-import lombok.Data;
-
-import java.time.LocalDateTime;
-import java.util.List;
-
-@Data
-public class ErrorResponse {
-    private String correlationId;
-    private LocalDateTime timeStamp;
-    private String code;
-    private String message;
-    private List<String> details;
-
-}
